requirements.txt
opencv-python~=4.8.0.74
mediapipe~=0.10.2
numpy~=1.24.3
tensorflow~=2.15.0
tqdm~=4.66.1
streamlit~=1.32.0
Pillow~=10.2.0